Added Delete Account

+++ Prayer Calendar/Views/Calendar View/PrayerNameInputView.swift
@@ -180,17 +180,24 @@ struct PrayerNameInputView: View {
             for usernameOrName in removals {
                 if usernameOrName.contains("/") == false { // This checks if the person is a linked account or not. If it was linked, usernameOrName would be a username. Usernames cannot have special characters in them.
                     
-                    let person = try await PrayerPersonHelper().retrieveUserInfoFromUsername(person: Person(username: usernameOrName), userHolder: userHolder) // retrieve the "person" structure based on their username. Will return back user info.
-                    
-                    // Update the friends list of the person who you have now removed from your list. Their friends list is updated, so that when they post, it will not add to your feed.
-                    let refFriends = db.collection("users").document(person.userID).collection("friendsList").document(userHolder.person.userID)
-                    try await refFriends.delete()
-                    
-                    // Update your prayer feed to remove that person's prayer requests from your current feed.
-                    let refDelete = try await db.collection("prayerFeed").document(userHolder.person.userID).collection("prayerRequests").whereField("userID", isEqualTo: person.userID).getDocuments()
-                    for document in refDelete.documents {
-                        try await document.reference.delete()
+                    if await PrayerPersonHelper().checkIfUsernameExists(username: usernameOrName) == true {
+                        do {
+                            let person = try await PrayerPersonHelper().retrieveUserInfoFromUsername(person: Person(username: usernameOrName), userHolder: userHolder) // retrieve the "person" structure based on their username. Will return back user info.
+                            
+                            // Update the friends list of the person who you have now removed from your list. Their friends list is updated, so that when they post, it will not add to your feed.
+                            let refFriends = db.collection("users").document(person.userID).collection("friendsList").document(userHolder.person.userID)
+                            try await refFriends.delete()
+                            
+                            // Update your prayer feed to remove that person's prayer requests from your current feed.
+                            let refDelete = try await db.collection("prayerFeed").document(userHolder.person.userID).collection("prayerRequests").whereField("userID", isEqualTo: person.userID).getDocuments()
+                            for document in refDelete.documents {
+                                try await document.reference.delete()
+                            }
+                        } catch {
+                            throw error
+                        }
                     }
+                    
                 } else { //else is for any names you have added which do not have a username; under your account and not linked.
                     
                     // Fetch all prayer requests with that person's first name and last name, so they are removed from your feed.
